requirements.txt
Flask==2.2.5
gunicorn==20.1.0
openai==1.3.0
flask-cors==4.0.0
waitress==2.1.2
python-dotenv==1.0.0
